Update n64-gameshark.md (#54)
v2.50 confirmed to be pristine as per the official Gameshark website
upgrade policy page:
https://web.archive.org/web/20010618223715/http://www.gameshark.com/articles/index.jsp?content_id=4878

"The current version number of the standard GameShark device is 2.5. The
current version of the GameShark Pro device is 3.3. Both of these
devices have 2,093 codes for 188 games and also includes the long enable
code for the Donkey Kong® 64 game. The GameShark Pro device contains
code creation utilities that allow you to create your own GameShark
codes."

Signed-off-by: Richard Weick <55462393+RWeick@users.noreply.github.com>